refactor: accept node for title and description
